pyproject.toml
[tool.poetry]
name = "startup-fin-os-mvp"
version = "0.1.0"
description = "AI-powered Startup Financial OS (MVP)"
authors = ["kophysty <kophysty@github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.35"
pydantic = "^2.7"
openai = "^1.30"
python-dotenv = "^1.0"
pandas = "^2.2"
pyyaml = "^6.0"
pdfkit = "^1.0"
sentence-transformers = "^2.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
responses = "^0.25"
black = "^24.4"
ruff = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 